Fix jemkins moving
