+++ tests/providers/jenkins/operators/test_jenkins_job_trigger.py
